refactor(tests)!: split example and DML-flow operations into schema and data migrations

+++ packages/zero-migrate-cli/hr-system/migrations/20260101000010_seed_reference_data.ts
@@ -44,7 +44,7 @@ export default {
         ]),
     });
     table("job_grades").delete({
-      where: (col) => col("id").in([int64("1"), int64("2"), int64("3"), int64("4")]),
+      where: (col) => col("id").in(["1", "2", "3", "4"]),
     });
     table("departments").delete({
       where: (col) =>
